fix autorepeat issue
Fixed (by workaround)
---------------------
To simulate auto repeating a key when holding down a key the OS does
something like that :

- send key-pressed
- wait until the auto-repeat starts (around a second)
- send key-release
REPEAT {
  - wait for about 50ms
  - send key-press followed by an
    immediate key-release
} UNTIL key is *really* realesed 

However, Java cannot check for this "real" releasing of the keyboard.

In result, this confuses some of the methods in GLTastatur. GLTastatur
used to regard a key as not being pressed in this 50ms cycle which leads
to stuttering e.g. in games with the arrow keys.

So now we only regard a key to not being pressed when 30ms have passed
after the latest key-release event. In practice this seems to be a
reasonable compromise between having a little later key-release and no
key-press at all.